refactor: add priority to event handling and improve query
