+++ api/event.go
@@ -1,15 +1,20 @@
 package api
 
 import (
+	"time"
+
 	"github.com/flanksource/duty/types"
 	"github.com/google/uuid"
 )
 
 type Event struct {
-	ID         uuid.UUID           `json:"id"`
-	Name       string              `json:"name"`
-	Properties types.JSONStringMap `json:"properties"`
-	Error      string              `json:"error"`
+	ID          uuid.UUID           `json:"id"`
+	Name        string              `json:"name"`
+	Properties  types.JSONStringMap `json:"properties"`
+	Error       string              `json:"error"`
+	Attempts    int                 `json:"attempts"`
+	LastAttempt *time.Time          `json:"last_attempt"`
+	Priority    int                 `json:"priority"`
 }
 
 // We are using the term `Event` as it represents an event in the
